+++ back-end/app/Http/Controllers/API/CabosController.php
@@ -24,14 +24,23 @@ class CabosController extends Controller
             'auto_eletrica_id' => 'required|integer'
         ]);
         if ($validator->fails())
-            return response()->json($validator->errors());
+            return response()->json([
+                'mensagem' => 'Preencha todos os campos',
+                $validator->errors()
+            ], 400);
 
         if (!AutoEletrica::find($request->auto_eletrica_id))
             return response()->json('Auto Eletrica nao existe');
 
-        $request->request->add(['user_name_created' => $request->user()->name]);
-        $cabo = Cabos::create($request->all());
-        return response()->json($cabo);
+        if ($request->user()->id <= 5) {
+            $request->request->add(['user_name_created' => $request->user()->name]);
+            $cabo = Cabos::create($request->all());
+            return response()->json($cabo);
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão para criar cabos'
+            ]);
+        }
     }
 
     public function update(Request $request, $id)
@@ -42,26 +51,47 @@ class CabosController extends Controller
             'quantidade' => 'integer'
         ]);
         if ($validator->fails())
-            return response()->json($validator->errors());
+            return response()->json([
+                'mensagem' => 'Preencha todos os campos',
+                $validator->errors()
+            ], 400);
 
-        $cabo = Cabos::find($id);
-        $request->request->add(['user_name_updated' => $request->user()->name]);
-        $cabo->update($request->all());
-        return response()->json($cabo);
+        if ($request->user()->id <= 5) {
+            $cabo = Cabos::find($id);
+            $request->request->add(['user_name_updated' => $request->user()->name]);
+            $cabo->update($request->all());
+            return response()->json($cabo);
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão para atualizar cabos'
+            ]);
+        }
     }
 
     public function changeSituation(Request $request, $id) {
         $cabo = Cabos::find($id);
 
-        $cabo->situacao = !$cabo->situacao;
-        $cabo->user_name_updated = $request->user()->name;
-        $cabo->update();
-        return response()->json($cabo);
+        if ($request->user()->id <= 5) {
+            $cabo->situacao = !$cabo->situacao;
+            $cabo->user_name_updated = $request->user()->name;
+            $cabo->update();
+            return response()->json($cabo);
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão'
+            ]);
+        }
     }
 
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
-        $cabo = Cabos::find($id);
-        $cabo->delete();
+        if ($request->user()->id <= 5) {
+            $cabo = Cabos::find($id);
+            $cabo->delete();
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão para deletar cabos'
+            ]);
+        }
     }
 }
