+++ back-end/app/Http/Controllers/API/ResponsavelController.php
@@ -23,14 +23,23 @@ class ResponsavelController extends Controller
             'auto_eletrica_id' => 'required|integer'
         ]);
         if ($validator->fails())
-            return response()->json($validator->errors());
+            return response()->json([
+                'mensagem' => 'Preencha todos os campos',
+                $validator->errors()
+            ], 400);
 
         if (!AutoEletrica::find($request->auto_eletrica_id))
             return response()->json('Auto Eletrica nao existe');
 
-        $request->request->add(['user_name_created' => $request->user()->name]);
-        $responsavel = Responsavel::create($request->all());
-        return response()->json($responsavel);
+        if ($request->user()->id <= 5) {
+            $request->request->add(['user_name_created' => $request->user()->name]);
+            $responsavel = Responsavel::create($request->all());
+            return response()->json($responsavel);
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão para adicionar responsáveis'
+            ]);
+        }
     }
 
     public function update(Request $request, $id)
@@ -40,17 +49,32 @@ class ResponsavelController extends Controller
             'telefone' =>  'string|max:20',
         ]);
         if ($validator->fails())
-            return response()->json($validator->errors());
+            return response()->json([
+                'mensagem' => 'Preencha todos os campos',
+                $validator->errors()
+            ], 400);
 
-        $responsavel = Responsavel::find($id);
-        $request->request->add(['user_name_updated' => $request->user()->name]);
-        $responsavel->update($request->all());
-        return response()->json($responsavel);
+        if ($request->user()->id <= 5) {
+            $responsavel = Responsavel::find($id);
+            $request->request->add(['user_name_updated' => $request->user()->name]);
+            $responsavel->update($request->all());
+            return response()->json($responsavel);
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão para atualizar responsáveis'
+            ]);
+        }
     }
 
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
-        $responsavel = Responsavel::find($id);
-        $responsavel->delete();
+        if ($request->user()->id <= 5) {
+            $responsavel = Responsavel::find($id);
+            $responsavel->delete();
+        } else {
+            return response()->json([
+                'mensagem' => 'você não tem permissão para deletar responsáveis'
+            ]);
+        }
     }
 }
